include/event: Expose HANDLED_CONT return value
The previous changes to the event hide if the last handler that
successfully handled the event allowed further event processing
(i.e., on success always UK_EVENT_HANDLED is returned). This
commit fixes this so that the caller of uk_raise_event() is informed
about if event processing can continue by other means or if it
should stop.

Signed-off-by: Marc Rittinghaus <marc.rittinghaus@unikraft.io>
Reviewed-by: Marco Schlumpp <marco@unikraft.io>
Approved-by: Marco Schlumpp <marco@unikraft.io>
Tested-by: Unikraft CI <monkey@unikraft.io>
GitHub-Closes: #738