lib/nolibc: Add -ffreestanding flag
GCC optimizes the use of `memcmp()` if the `-ffreestanding` flag is
missing. In the current configuration, the flag is missing for GCC <= 9.
The resulting optimization causes in an infinite loop for `lib/nolibc`.

This commit fixes the issue by adding the `-ffreestanding` flag for
`lib/nolibc`, irrespective of the GCC version.

GitHub-Fixes: #728

Signed-off-by: Razvan Deaconescu <razvand@unikraft.io>