lib/vfscore: Cast initrd vbase to (void *)
PR #722 (Boot code refactoring) introduced an update to
`struct ukplat_memregion_desc` where the virtual base of a memory region is
now specified as a __vaddr_t integer value. Since it is a virtual address
we can cast it to (void *) to avoid build warnings in
`lib/vfscore/rootfs.c`.

Signed-off-by: Simon Kuenzer <simon@unikraft.io>