lib/uksched: Align allocated stacks to arch requirement
This commits introduces aligns stack allocation that are done by
`lib/uksched` to the requirement of the target architecture. Without
aligned stacks, random crashed may occur because the compiler is assuming
such an alignment. An incorrect stack alignment can cause crashes because
the compiler may placed instructions in the code that will fail when
accessing values on the stack if not properly aligned.

Signed-off-by: Simon Kuenzer <simon@unikraft.io>