plat/common/x86: 16-byte align syscall stack
Currently, the user stack is used as kernel stack. This can be
a problem if the stack is not aligned on syscall entry. However, it
may be aligned for some calls. We thus cannot just push a fixed
amount of words. Instead, we store the original stack pointer and
align it using an and operation, which guarantees correct alignment
in any case.

Signed-off-by: Marc Rittinghaus <marc.rittinghaus@unikraft.io>