lib/netdev: Make netdev SMP safe
This commit synchronizes the operations which modify the network device.
These involves network configuration and setters.
The network device layer assumes that the queue operations are
synchronized by the lower level such as virtio, hence it does not
acquire the lock for such operations. This allows us to parallelize the
queue operations.

Signed-off-by: Sairaj Kodilkar <skodilkar7@gmail.com>