[WILL BE REMOVED] lib/uklock/rwlock: Separate the owner field and flags
Previously the owner and flags for the reader writer lock were store in
a single variable. But it required that the stack is aligned to the
stack length. Since new changes to the stack aligns it to only 16 bytes,
this implementation does not work and is buggy.
Hence this commit separates the owner and flag field. Instead of stack
bottom, new implementation uses pointer to the `struct uk_thread` as the
owner of the lock

Signed-off-by: Sairaj Kodilkar <skodilkar7@gmail.com>