[WILL BE REMOVED] lib/uklock/rwlock: Fix lost wakeup problem
Previously the lock simply set the waiters flag, but this can create
lost wakeup problem when unlock and wakeup has completed before the
thread goes to sleep.
Hence new implementation uses atomic compare exchange. If we fail to set
the waiters flags, then we again check if the lock has been released by
the owner or not.

Checkpatch-Ignore: LONG_LINE
Signed-off-by: Sairaj Kodilkar <skodilkar7@gmail.com>