lib/posix-mmap: Fix wrong errno value
The currently used syscall shim macro handles a return value of -1
by setting the return value to the value of errno, which results in
incorrect behaviour with mmap, which returns a MAP_FAILED (-1) address
as an indicator of failure. The result is that in case of mmap error,
the syscall will return the value of errno as the address, rather
than returning -1 and setting errno.

This commit fixes the issue by setting errno to the correct value
in case of mmap error.

Co-authored-by: Razvan Deaconescu <razvand@unikraft.io>
Co-authored-by: Stefan Jumarea <stefanjumarea02@gmail.com>
Signed-off-by: Ioan-Teodor Teugea <ioan_teodor.teugea@stud.acs.upb.ro>
Reviewed-by: Stefan Jumarea <stefanjumarea02@gmail.com>
Approved-by: Marc Rittinghaus <marc.rittinghaus@unikraft.io>
Tested-by: Unikraft CI <monkey@unikraft.io>
GitHub-Closes: #799