plat: Fix DTB interrupt node lacking address-cells property
The `gen_pci_irq_parse` function fails when an interrupt parent node is
missing the `#address-cells` property. Even though the Devicetree spec
states that this property shall be explicitly defined, some DTBs
(such as the RISC-V DTB generated by QEMU <= 7.1) do not include this
property for the interrupt controller. However, when the parent
unit address is missing from the DTB, we can assume that the parent
node's `#address-size` property is equal to 0 to maintain compatibility
with such DTBs. The specification even says that "If a unit address
component is not required, `#address-cells` shall be explicitly
defined to be zero.".

We can thus introduce a new CONFIG option,
`DTB_INTERRUPT_ADDRESS_CELLS_ERRATA`, which should provide an errata for
uncomplying device trees when parsing the interrupt nodes.

Signed-off-by: Eduard Vintilă <eduard.vintila47@gmail.com>