plat/drivers/ofw: Fix search in `fdt_node_offset_by_compatible_list`
The current implementation returns as soon as a node that matches a
compatible string is found. However, we should iterate through all of
the compatible list to make sure that we return the node with the least
offset value such that future searches will not skip any matching nodes.

Signed-off-by: Eduard Vintilă <eduard.vintila47@gmail.com>
Reviewed-by: Robert Kuban <robert.kuban@opensynergy.com>
Reviewed-by: Razvan Virtan <virtanrazvan@gmail.com>
Approved-by: Michalis Pappas <michalis@unikraft.io>
Tested-by: Unikraft CI <monkey@unikraft.io>
GitHub-Closes: #805