lib/ukargparse: Support escaping of quotes
This commit introduces the support of escaping single (`'`) and double
quotes (`"`) with the escape character `\`. The behaviour is similar to
what one expects from Unix shells, like `sh` or `bash`.

Single-quoted characters and character sequences preserve each character.
Double-quoted characters and character sequences preserve each character
except the backslash `\`. The backslash can be used to switch of the
special meaning of a character, like the double quote (`\"`) or a
backslash (`\\`).

Signed-off-by: Simon Kuenzer <simon@unikraft.io>
Reviewed-by: Razvan Deaconescu <razvand@unikraft.io>
Approved-by: Razvan Deaconescu <razvand@unikraft.io>
Tested-by: Unikraft CI <monkey@unikraft.io>
GitHub-Closes: #893