lib/ukdebug: Add crashdump printing API
The crashdump API is inspired by the asmdump and hexdump
functions and allows printing a register dump, stack dump,
and a function call trace based on a supplied register
snapshot. All memory accesses are checked with uknofault
if available so dumps are less likely to crash the system
for a corrupted execution state.

Co-authored-by: Marc Rittinghaus <marc.rittinghaus@kit.edu>
Signed-off-by: Marco Schlumpp <marco@unikraft.io>