build: Do not use segment annotations on linuxu
For the linuxu platform we rely on the default linker script and only
add some additional sections on top of that one. That also means that
there aren't any segments defined. This makes the segment specification
opt-in instead of assuming the segments are always defined.

Signed-off-by: Marco Schlumpp <marco@unikraft.io>
Checkpatch-Ignore: SPACING
Checkpatch-Ignore: COMPLEX_MACRO