lib/9pfs: Add truncate to 9P_PROTO_2000U
When providing a length in the stat structure the 9pfs server on the
host will truncate the file to the specified size. We use this to also
enable truncation for the 2000U protocol version. Since uk_9pfs_setattr
checks for the protocol to use, the check can be completely dropped in
uk_9pfs_truncate.

Signed-off-by: Marc Rittinghaus <marc.rittinghaus@unikraft.io>