plat: Generic clock event framework
This introduces the ukplat_clock_event data structure which abstracts
over programmable timers. This is intended to be a primitive for
higher-level libraries and as such this only allows a single client for
a timer.
It is currently limited to non-periodic timers but can be extended to
also support periodic timers in the future.

Signed-off-by: Marco Schlumpp <marco@unikraft.io>