plat/x86: Register the APIC timer to clock_event system
There is one LAPIC timer for each core in a modern x86 system. Compared
to the legacy PIT timer, it is faster to program. A slightly more recent
addition is the TSC-deadline mode, which allows generating an interrupt
once the TSC reaches a specific values. Compared to the classic one-shot
timer mode, we don't need to find out the frequency of the timer and can
directly use TSC-based timestamps.

Signed-off-by: Marco Schlumpp <marco@unikraft.io>