lib/uktimer: Add timer-wheel-based timers
The design of the timer-wheel is based off the one currently used by the
Linux kernel. Compared to the high-resolution timers, this is optimized
for timeouts which are usually removed before they would fire.
Adding/removing timers from the timer wheel is a simple list `O(1)`
add/remove operation instead of `O(log n)` with hr-timers.

Signed-off-by: Marco Schlumpp <marco@unikraft.io>