lib/vfscore: Make vnops strings constant
The current vnode operations define strings as mutable, which
makes it more difficult to do optimizations. Since no buffer size
is supplied the strings can only be modified on a char-for-char
basis anyways, which also limits the use-cases. In addition, no
implementation of the vnode operations makes use of the
mutability of the strings.

This commit thus introduces the `const` keyword for all strings
passed to vnode operations.

Signed-off-by: Marc Rittinghaus <marc.rittinghaus@unikraft.io>