lib/9pfs: Fix wrong fid for fsync in 9P2000.u
Performing an fsync with the 9P2000.u protocol version previously
resulted in an I/O error. This is because we used the wrong fid
in the request. This commit moves the implementation of setattr
into a separate function that accepts the fid to use, and adjusts all
callers to provide the fid.

Signed-off-by: Marc Rittinghaus <marc.rittinghaus@unikraft.io>
Reviewed-by: Stefan Jumarea <stefanjumarea02@gmail.com>
Reviewed-by: Eduard Vintilă <eduard.vintila47@gmail.com>
Approved-by: Razvan Deaconescu <razvand@unikraft.io>
Tested-by: Unikraft CI <monkey@unikraft.io>
GitHub-Closes: #886