lib/vfscore: Avoid recursive locking in sys_rename
If the source and destination directories are the same in sys_rename,
the vnode is locked two times. This is problematic with non-recursive
mutexes. Also the double checks for access are not needed. This commit
adds a check to make sure in this case, only a single lock is done and
redundant checks are avoided.

Signed-off-by: Marc Rittinghaus <marc.rittinghaus@unikraft.io>