plat/common: Implement insert method for the legacy x86 high memory
Write a short inline function to insert the x86 legacy high memory
area right below the first 1MiB, in the `0xA0000` to `0x100000` range.

Preferably, since this area is very low in memory, it should be added
before we start building the memory region list to save some iterations.

Signed-off-by: Sergiu Moga <sergiu.moga@protonmail.com>
Reviewed-by: Michalis Pappas <michalis@unikraft.io>
Approved-by: Razvan Deaconescu <razvand@unikraft.io>
Tested-by: Unikraft CI <monkey@unikraft.io>
GitHub-Closes: #848