plat/drivers/gic: Enable `GICv2` probing through `ACPI`
If `ACPI` is enabled, rely on it for probing the `GICv2`. Thus, we
need to take two steps for this:
- fetch the GIC CPU interface's base address and default length from
the `GICC` `MADT` entry. Since there is only one `Redistributor`, all
`GICC`'s must have the same physical base address.
- fetch the `GIC Distributor`'s physical base
address from `MADT`'s `GICD` structure and check that the size is
compliant with that of `GICv2`.

`GICR Base Address` must be set to 0, otherwise this is `GICv3/4`.

Furthermore, separate the two ways of probing `GICv2` in two:
one for `ACPI` and one for `Devicetree`.

Signed-off-by: Sergiu Moga <sergiu.moga@protonmail.com>