lib/ukallocbuddy: Add debug freelist sanity checks
This adds a Kconfig option to enable sanity checking of buddy allocator
free lists at runtime. Check points are placed at the beginning and end
of functions that operate on the free lists. When the option is disabled
the checks default to zero-overhead no-ops.
In addition, this adds an assert that returned allocated pages are
correctly aligned to their buddy order.

Signed-off-by: Andrei Tatar <andrei@unikraft.io>