plat/common: Ensure `IRQ`'s disabled in `ukplat_lcpu_halt_to`
Make sure that `IRQ`'s are disabled when calling `ukplat_lcpu_halt_to`.
Furthermore, do not surround the call to `time_block_until` with `IRQ`
`save`/`restore` and inform the function user through a `NOTE` to take
care of that himself.

Therefore, with this commit, change every other reference to this
function accordingly. Furthermore, change the name of the function to
reflect this.

Signed-off-by: Sergiu Moga <sergiu@unikraft.io>
Suggested-by: Marco Schlumpp <marco@unikraft.io>