plat/kvm/arm: Do not enable interrupts before `wfi`
There is the opportunity to miss an IRQ on the path to `wfi` from the
idle thread after deciding to sleep:
The interrupts are enable just before exuting `wfi`.
If the interrupt already arrived before that, it will be executed
before `wfi`, and `wfi` will go to sleep afterward, not being woken
up by the interrupt.

This patch removes calls to enable/disable interrupts around `wfi`.
`wfi` will be woken up by pending interrupts regardless.

Signed-off-by: Robert Kuban <robert.kuban@opensynergy.com>